+++ backend/models/index.js
@@ -6,4 +6,4 @@ mongoose.connect(process.env.MONGO_URI, {
   useUnifiedTopology: true,
 });
 
-module.exports.Movie = require("../models/movie");
+module.exports.Movie = require("./movieModel");
