+++ backend/models/movieModel.js
@@ -1,18 +1,6 @@
 const mongoose = require("mongoose");
 const { Schema } = mongoose;
 
-// Schema
-// const movieSchema = new Schema({
-//   movie_name: { type: String, required: true },
-//   release_year: { type: Number },
-//   genre: { type: String },
-//   runtime_minutes: { type: Number },
-//   total_revenue: { type: String },
-//   main_actor: { type: String },
-//   image: { type: String },
-//   description: { type: String },
-// });
-
 const movieSchema = new Schema({
   movie_name: {
     type: String,
